Add files via upload
The following code is representative model of Google's search engine. On the Google Search page, the user will be able to type in a query, click “Google Search”, and will taken to the Google search results for that page. The page also has links on the top-right6 corner, to navigate to "Google Images" & "Advanced Search."
An "I am feeling lucky" button has also been added next to the search button. Searching with the help of the “I’m Feeling Lucky” button on the main Google Search page, will direct the user to the first Google search result for the query, bypassing the normal search results page.