Add files via upload
This piece of code is intended to be a web-page functionally similar to that of Google Images. The text entered in the search box will take users to the image results of their input. Icons at the top right-hand side of the page will take the user/s back to Google Home, or Google Advanced Search, whichever they want. Lastly, the "I am Feeling Lucky" button has been added too, with the same role as that which you can find in the Home Page of "Google" (read, index.html code in the same repo)!